% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{check_and_download}
\alias{check_and_download}
\title{Check for HISTALP data on system}
\usage{
check_and_download(base_dir = "~")
}
\arguments{
\item{base_dir}{where to store the HISTALP grid files}
}
\value{
nothing, invoked for side effects
}
\description{
Check, if HISTALP data is already present on the system, if not download and
install locally.
}
\examples{
\dontrun{
check_and_down
}
}

