% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histalp.R
\name{get_histalp}
\alias{get_histalp}
\title{Extract climate data from nearest HISTALP gridcell}
\usage{
get_histalp(lon, lat, na.rm = TRUE, base_dir = "~")
}
\arguments{
\item{lon}{Longitude}

\item{lat}{Latitude}

\item{na.rm}{Should the output be restricted to the common time span for all
variables (default is TRUE)}

\item{base_dir}{Where to look for the HISTALP netCDF files}
}
\value{
a \code{data.frame}
}
\description{
Extract climate data from nearest HISTALP gridcell
}
\examples{
get_histalp(10.2, 44.8)
}
